move resistances to own file
